--break Abstract Object Attr/Object Attr/Stage Attr into separate files.
