+++ src/esp/metadata/managers/AbstractObjectAttributesManagerBase.h
@@ -10,9 +10,9 @@
  * esp::metadata::managers::AbstractObjectAttributesManager
  */
 
-#include "esp/metadata/attributes/ObjectAttributes.h"
-
 #include "esp/metadata/MetadataUtils.h"
+#include "esp/metadata/attributes/AbstractObjectAttributes.h"
+#include "esp/metadata/managers/AssetAttributesManager.h"
 #include "esp/metadata/managers/AttributesManagerBase.h"
 
 namespace Cr = Corrade;
